refactor(pdf): isolate equation transcripts
